refactor: move it to eval
